+++ tests/Persistence/EntityPersisterTest.php
@@ -5,17 +5,23 @@ namespace Somnambulist\Components\ApiClient\Tests\Persistence;
 use IlluminateAgnostic\Str\Support\Str;
 use PHPUnit\Framework\TestCase;
 use Somnambulist\Collection\Contracts\Collection;
-use Somnambulist\Components\ApiClient\Client\Connection;
 use Somnambulist\Components\ApiClient\Client\ApiRoute;
 use Somnambulist\Components\ApiClient\Client\ApiRouter;
-use Somnambulist\Components\ApiClient\Client\ApiService;
+use Somnambulist\Components\ApiClient\Client\Connection;
+use Somnambulist\Components\ApiClient\Manager;
 use Somnambulist\Components\ApiClient\Persistence\Actions\CreateAction;
 use Somnambulist\Components\ApiClient\Persistence\Actions\DestroyAction;
 use Somnambulist\Components\ApiClient\Persistence\Actions\UpdateAction;
 use Somnambulist\Components\ApiClient\Persistence\EntityPersister;
 use Somnambulist\Components\ApiClient\Persistence\Exceptions\EntityPersisterException;
-use Somnambulist\Components\ApiClient\Tests\Stubs\Entities\User;
 use Somnambulist\Components\ApiClient\Tests\Support\Behaviours\UseFactory;
+use Somnambulist\Components\ApiClient\Tests\Support\Stubs\Entities\User;
+use Somnambulist\Components\AttributeModel\TypeCasters;
+use Somnambulist\Domain\Entities\Types\Geography\Country;
+use Somnambulist\Domain\Entities\Types\Identity\EmailAddress;
+use Somnambulist\Domain\Entities\Types\Identity\Uuid;
+use Somnambulist\Domain\Entities\Types\PhoneNumber;
+use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\HttpClient\MockHttpClient;
 use Symfony\Component\HttpClient\Response\MockResponse;
 use Symfony\Component\Routing\RouteCollection;
@@ -37,33 +43,30 @@ class EntityPersisterTest extends TestCase
 
     use UseFactory;
 
-    /**
-     * @var \Somnambulist\Components\ApiClient\Persistence\EntityPersister
-     */
-    private $persister;
+    private ?EntityPersister $persister = null;
 
     protected function setUp(): void
     {
         $host    = 'http://api.example.dev/users/v1';
-        $new     = new MockResponse(file_get_contents(__DIR__ . '/Stubs/user_store.json'), [
+        $new     = new MockResponse(file_get_contents(__DIR__ . '/../Support/Stubs/json/user_store.json'), [
             'http_code'        => 201,
             'response_headers' => [
                 'Content-Type' => 'application/json',
             ],
         ]);
-        $updated = new MockResponse(file_get_contents(__DIR__ . '/Stubs/user_updated.json'), [
+        $updated = new MockResponse(file_get_contents(__DIR__ . '/../Support/Stubs/json/user_updated.json'), [
             'http_code'        => 200,
             'response_headers' => [
                 'Content-Type' => 'application/json',
             ],
         ]);
-        $deleted = new MockResponse(file_get_contents(__DIR__ . '/Stubs/user_deleted.json'), [
+        $deleted = new MockResponse(file_get_contents(__DIR__ . '/../Support/Stubs/json/user_deleted.json'), [
             'http_code'        => 204,
             'response_headers' => [
                 'Content-Type' => 'application/json',
             ],
         ]);
-        $error   = new MockResponse(file_get_contents(__DIR__ . '/Stubs/user_error.json'), [
+        $error   = new MockResponse(file_get_contents(__DIR__ . '/../Support/Stubs/json/user_error.json'), [
             'http_code'        => 400,
             'response_headers' => [
                 'Content-Type' => 'application/json',
@@ -89,14 +92,22 @@ class EntityPersisterTest extends TestCase
         };
         $client   = new MockHttpClient($callback);
 
-        $router = new ApiRouter(new ApiService($host, 'users'), new RouteCollection());
+        $router = new ApiRouter($host, new RouteCollection());
         $router->routes()->add('users.create', new ApiRoute('/users', [], ['POST']));
         $router->routes()->add('users.update', new ApiRoute('/users/{id}', ['id' => '[0-9a-f\-]{36}'], ['PUT', 'PATCH']));
         $router->routes()->add('users.destroy', new ApiRoute('/users/{id}', ['id' => '[0-9a-f\-]{36}'], ['DELETE']));
 
-        $client = new Connection($client, $router);
+        $client = new Connection($client, $router, new EventDispatcher());
 
-        $this->persister = new EntityPersister($client, $this->factory()->makeUserMapper());
+        new Manager(['default' => $client], [
+            new TypeCasters\DateTimeCaster(),
+            new TypeCasters\SimpleValueObjectCaster(Uuid::class, ['uuid']),
+            new TypeCasters\SimpleValueObjectCaster(EmailAddress::class, ['email']),
+            new TypeCasters\SimpleValueObjectCaster(PhoneNumber::class, ['phone']),
+            new TypeCasters\EnumerableKeyCaster(Country::class, ['country']),
+        ]);
+
+        $this->persister = new EntityPersister($client);
     }
 
     protected function tearDown(): void
@@ -126,10 +137,10 @@ class EntityPersisterTest extends TestCase
     {
         $this->expectException(EntityPersisterException::class);
         $this->expectExceptionCode(422);
-        $this->expectExceptionMessage('Entity of type "Somnambulist\Components\ApiClient\Tests\Stubs\Entities\User" could not be created');
+        $this->expectExceptionMessage(sprintf('Entity of type "%s" could not be created', User::class));
 
         $repo = $this->persister;
-        $req = CreateAction::new(User::class)
+        $req  = CreateAction::new(User::class)
             ->with([
                 'name' => 'foo bar', 'email' => 'foo@example.com', 'error' => true,
             ])
@@ -177,7 +188,7 @@ class EntityPersisterTest extends TestCase
 
         $this->assertInstanceOf(User::class, $user);
         $this->assertEquals('c8259b3b-8603-3098-8361-425325078c9a', $user->id->toString());
-        $this->assertTrue($user->active);
+        $this->assertTrue($user->is_active);
     }
 
     public function testDestroy()